Use helper class to increase reusability
