+++ Stack/Problems/InfixToPostFixConverter/InfixToPostfixConverter.hpp
@@ -1,12 +1,7 @@
 #include <string>
 #include <unordered_map>
-#include <stdexcept>
+#include "../helper.hpp"
 #include "../../Linked List Stack/Stack.hpp"
-
-class InvalidExpression : public std::invalid_argument {
-  public:
-    InvalidExpression(std::string message="") : std::invalid_argument(message) {}
-};
 
 /*
   Tells whether operator 1 has precedence over operator 2
@@ -32,22 +27,6 @@ inline bool _hasPrecedence(char op1, char op2) {
 // Tell wheter given character is an operator or not
 inline bool _isOperator(const char& ch) {
   return ch == '+' || ch == '-' || ch == '*' || ch == '/';
-}
-
-inline int _findNextNonDigitIndex(std::string str, int startingPos) {
-
-  if (startingPos > str.size()) throw std::out_of_range("Invalid starting position given!"); 
-
-  int nextNonDigitIndex;
-
-  for (nextNonDigitIndex = startingPos; nextNonDigitIndex < str.size(); nextNonDigitIndex++) {
-    if (str[nextNonDigitIndex] < '0' || str[nextNonDigitIndex] > '9') {
-      return nextNonDigitIndex;
-     }
-  }
-
-  // If there is no next non digit index, return -1
-  return -1;
 }
 
 // Converts given infix form to postfix form
@@ -86,7 +65,7 @@ inline std::string convertToPostfix(const std::string& infixExpr) {
     // If current index is at digit (a potential number),
     // extract the number and start from next index after the number 
     else if (std::isdigit(infixExpr[i])) {
-       int nextNonDigitIndex = _findNextNonDigitIndex(infixExpr, i);
+       int nextNonDigitIndex = Helper::findNextNonDigitIndex(infixExpr, i);
 
       postfixExpr += infixExpr.substr(i, nextNonDigitIndex == -1? std::string::npos : nextNonDigitIndex - i) + ' ';
